backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
twilio==9.2.3
gunicorn==22.0.0


